Move string formatting utils to own module
